remove unused code and refactor maf validation
