+++ tests/test_maf.py
@@ -197,41 +197,3 @@ def test_validate_that_allele_freq_are_not_na_does_expected_logging(
         )
     else:
         assert len(caplog.records) == 0
-
-
-@pytest.mark.parametrize(
-    "df, expect_error, error",
-    [
-        # Case 1: No missing columns
-        (
-            pd.DataFrame(
-                {col: ["dummy"] for col in maf_to_cbio.REQUIRED_MAF_COLS},
-            ),
-            False,
-            ""
-        ),
-        # Case 2: Has missing column
-        (
-            pd.DataFrame(
-                {
-                    col: ["dummy"]
-                    for col in maf_to_cbio.REQUIRED_MAF_COLS
-                    if col not in ["Annotation_Status"]
-                },
-            ),
-            True,
-            "Missing required columns in maf: ['Annotation_Status']",
-        ),
-    ],
-    ids=["no_missing_cols", "has_missing_col"],
-)
-def test_validate_that_required_columns_are_present_expected_logging(
-    df, expect_error, error, caplog
-):
-    with caplog.at_level(logging.ERROR):
-        maf_to_cbio.validate_that_required_columns_are_present(df)
-
-    if expect_error:
-        assert len(caplog.records) == 1 and caplog.records[0].message == error
-    else:
-        assert len(caplog.records) == 0
